requirements.txt
python-dateutil==2.8.0
yaap>=0.3
tqdm==4.36.1
h5py==2.10.0
torch==1.3.0
numpy==1.16
spacy==2.2.1
gensim==3.8.1
pyyaml==5.1.2
pytorch-models==0.2.8
prettytable==0.7.2
requests==2.22.0
six==1.12.0
protobuf==3.10.0
stanza==0.3
vocab==0.0.3
embeddings==0.0.4
jsondiff==1.2.0
scipy==1.3.1
inflect==2.1.0
tensorboard==2.0.1
future==0.18.1
nltk==3.4.5
Pillow==6.2.1
rouge==0.3.2